refactor(review): 拆分 balance_query_provider.dart (652行) — 状态模型抽到 part
- balance_query_state.dart：BalanceResourceState/BalanceTrendState/RoomBinding
  （UI 快照）+ _ResourceEntry/_BalanceCacheKey/_TrendCacheKey/_TrendEntry（内部缓存）
- balance_query_provider.dart：仅保留 Provider 逻辑（绑定管理/加载/采样/趋势）
- 私有类经 part 共享，外部 import 零改动
